@86epcnmjw #Create navigationtype file and isolate Main styled component
