+++ packages/core-library/components/GenericDrawerLayout/content/Main.tsx
@@ -1,0 +1,27 @@
+import { styled } from '@mui/material';
+
+export const Main = styled('main', {
+    shouldForwardProp: (prop) => {
+        return !['isMobile', 'open'].includes(prop as string)
+    }
+})<{
+    open?: boolean;
+    isMobile?: boolean;
+}>(({ theme, open, isMobile }) => ({
+    flexGrow: 1,
+    transition: theme.transitions.create('margin', {
+        easing: theme.transitions.easing.sharp,
+        duration: theme.transitions.duration.leavingScreen,
+    }),
+    marginLeft: `-240px`,
+    ...(isMobile && {
+        marginLeft: 0,
+    }),
+    ...(open && {
+        transition: theme.transitions.create('margin', {
+            easing: theme.transitions.easing.easeOut,
+            duration: theme.transitions.duration.enteringScreen,
+        }),
+        marginLeft: 0,
+    }),
+}));
