+++ packages/core-library/components/index.ts
@@ -23,3 +23,5 @@ export * from "./blocks/LoginFormBlock/LoginFormBlock";
 export * from "./blocks/ReusableFooter/Footer";
 
 export * from "./DataGrid/DataGrid";
+export * from "./GenericDrawerLayout/DrawerLayout"
+export * from "./GenericSidebar/Sidebar"
