+++ packages/core-library/types/navigation.ts
@@ -1,0 +1,10 @@
+export type NavigationItemType = {
+    id: number
+    label: string
+    path?: string
+    icon?: React.ReactNode
+}
+
+export type NavigationType = NavigationItemType & {
+    children?: NavigationItemType[]
+}
